detectron/requirements.txt
Flask
opencv-python
requests
matplotlib
Pillow
torch
numpy
roboflow